+++ src/config.ts
@@ -9,22 +9,32 @@ interface Config {
     };
     timer: {
         pauseWhenUnfocused: boolean;
+        unfocusedThreshold: number;
+        pauseWhenIdle: boolean;
         idleThreshold: number;
     };
 }
 
 export function get_config(): Config {
     const config = vscode.workspace.getConfiguration('project-timer');
+    if (config.get("timer.unfocusedThreshold") && config.get("timer.unfocusedThreshold") as number < 0) {
+        console.warn(`Invalid value for 'project-timer.timer.unfocusedThreshold': ${config.get("timer.unfocusedThreshold")}. Must be a non-negative number.`);
+    }
+    if (config.get("timer.idleThreshold") && config.get("timer.idleThreshold") as number < 0) {
+        console.warn(`Invalid value for 'project-timer.timer.idleThreshold': ${config.get("timer.idleThreshold")}. Must be a non-negative number.`);
+    }
     return {
         statusBar: {
             enabled: config.get("statusBar.enabled", true) as Config['statusBar']['enabled'],
-            displayPrecision: config.get("statusBar.displayPrecision", "minute") as Config['statusBar']['displayPrecision'],
+            displayPrecision: config.get("statusBar.displayPrecision", "auto") as Config['statusBar']['displayPrecision'],
             displayProjectName: config.get("statusBar.displayProjectName", true) as Config['statusBar']['displayProjectName'],
-            displayToday: config.get("statusBar.displayToday", true) as Config['statusBar']['displayToday']
+            displayToday: config.get("statusBar.displayToday", false) as Config['statusBar']['displayToday']
         },
         timer: {
-            pauseWhenUnfocused: config.get("timer.pauseWhenUnfocused", false) as Config['timer']['pauseWhenUnfocused'],
-            idleThreshold: config.get("timer.idleThreshold", 300) as Config['timer']['idleThreshold']
+            pauseWhenUnfocused: config.get("timer.pauseWhenUnfocused", true) as Config['timer']['pauseWhenUnfocused'],
+            unfocusedThreshold: config.get("timer.unfocusedThreshold", 5) as Config['timer']['unfocusedThreshold'],
+            pauseWhenIdle: config.get("timer.pauseWhenIdle", false) as Config['timer']['pauseWhenIdle'],
+            idleThreshold: config.get("timer.idleThreshold", 5) as Config['timer']['idleThreshold']
         }
     };
 }